added simple query and download
a simple querying and download component with basic stora and query and
download services.

query result are shown below query field as v-cards. A batch download
button is shown at the top and each card has to buttons. One button for
single file download and one button to expand and view details.

the store commits only basic actions and mutations. a lot of the logging
functionality is still missing.

overall a lot of the design is still to be computed.