basic upload functionality
upload can be handled by drag-n-drop or by file selection
upload can be done in batch or single file mode

styling is still awful and needs fixing. card components are not showing
as intended.

store is not logging enough and file iinfos are not saved for later sessions in frontend.

buttons for retry of warning files are still missing

basic functionality:
components function commits store action. This action in turn logs
commits a logging mutation and then activates a service function that
sends the request. The store then commits a success/error mutation and
passes the data back to the component.