fixed upload colors and error handling
colors now work according to conformity check results:
black - fatal
red - error
yellow - warning
green - successful

errors are thrown if response is not ok and not status 406 (not
acceptable). Not acceptable is a response returned when conformity
checks fail and ignore_warnings/ignore_errors are not activated. The
status will be shown in the list of upload files and should not cause an
error.

Bugfixes:
- a naming error in authentication_helper caused credentials to be false